+++ src/ast/statement.rs
@@ -15,8 +15,8 @@ impl StatementListNode {
 }
 
 impl ASTNode for StatementListNode {
-    fn eval(&self, symtab: &mut HashMap<String, i32>) -> Result<i32, InterpreterError> {
-        let mut value: i32 = 0;
+    fn eval(&self, symtab: &mut HashMap<String, f64>) -> Result<f64, InterpreterError> {
+        let mut value: f64 = 0.;
 
         for node in self.nodes.iter() {
             value = node.eval(symtab)?;
